Add delay to wait_for task that checks for host after reboot
Without this delay, the system has not yet rebooted when the wait_for task fires. This
meant the playbook would contiune to run with a reboot pending. The host would
then reboot in the middle of a subsequent task, which happened to be yum. Rebooting
in the middle of the yum task creates inconsistencies in the rpm database, causing
subsequest executions of the playbook to fail on the yum task.